Peer Subscribtion: pseudo State Machine for Connection refactor
The problem with Connection tackeled in the commit is that Connection
was both Optional and had internal state that determined actual
readiness to work. That caused problems with providing a well defined
interface for sending out Packets. It required repetitive Some/None
matching and is_connected checks.

In general it should be either one of the two. The proposed
in the commit approach is a pseudo state machine that manages
actual connection state and allows to remove redundant Optionalities
of the Connection itself and of it's internal structure.